@@ -3,24 +3,25 @@ description: A new Flutter package project.
 version: 0.1.0
 publish_to: none
 environment:
-  sdk: ">=2.19.2 <3.0.0"
+  sdk: ^3.5.4
 
 dependencies:
   flutter:
     sdk: flutter
   bloc: ^8.0.2
-  flutter_reactive_ble: ^5.0.3
+  flutter_reactive_ble: ^5.3.1
   flutter_bloc: ^8.0.1
-  permission_handler: ^10.0.0
+  permission_handler: ^11.3.1
   vector_math: ^2.1.0
-  wakelock: ^0.6.1
   path_provider: ^2.0.5
-  uuid: ^3.0.6
+  uuid: ^4.5.1
+  rxdart: ^0.28.0
+  csv: ^6.0.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
-  dependency_validator: ^3.2.2
+  flutter_lints: ^5.0.0
+  dependency_validator: ^4.1.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
