@@ -1,4 +1,4 @@
-library flutter_remo;
+library;
 
 export 'src/bloc/bluetooth/bluetooth_bloc.dart';
 export 'src/bloc/remo/remo_bloc.dart';
