@@ -4,8 +4,9 @@ abstract class RemoEvent {}
 
 /// Starts connection to a device.
 class OnConnectDevice extends RemoEvent {
-  OnConnectDevice(this.address);
+  OnConnectDevice(this.address, this.name);
   final String address;
+  final String name;
 }
 
 class OnSwitchTransmissionMode extends RemoEvent {}
