@@ -4,9 +4,10 @@ part of 'remo_file_bloc.dart';
 abstract class RemoFileEvent {}
 
 class StartRecording extends RemoFileEvent {
-  final Stream<RemoData> remoDataStream;
+  final Stream<RmsData> rmsDataStream;
+  final Stream<ImuData> imuDataStream;
 
-  StartRecording(this.remoDataStream);
+  StartRecording(this.rmsDataStream, this.imuDataStream);
 }
 
 class StopRecording extends RemoFileEvent {}
@@ -14,6 +15,11 @@ class StopRecording extends RemoFileEvent {}
 class SaveRecord extends RemoFileEvent {
   SaveRecord(this.fileName);
   final String fileName;
+}
+
+class OpenRmsRecord extends RemoFileEvent {
+  OpenRmsRecord(this.filePath);
+  final String filePath;
 }
 
 class DiscardRecord extends RemoFileEvent {}
